Product Variant Items Crud and Product Delete, Status Update

+++ app/Http/Controllers/Backend/ProductVariantController.php
@@ -6,6 +6,7 @@ use App\DataTables\ProductVariantDataTable;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\ProductVariantRequest;
 use App\Http\Requests\ProductVariantUpdateRequest;
+use App\Interfaces\ProductVariantItemRepositoryInterface;
 use App\Interfaces\ProductVariantRepositoryInterface;
 use App\Models\Product;
 use Illuminate\Http\Request;
@@ -76,6 +77,13 @@ class ProductVariantController extends Controller
      */
     public function destroy(string $id)
     {
+        $productVariantItemCount = app(ProductVariantItemRepositoryInterface::class)->getVariantItemCountByVariant($id);
+
+        if($productVariantItemCount > 0)
+        {
+            return response(['status' =>'error' , 'message' =>"This Variant contains variant items. Inorder to delete this Variant you have to delete all variant items first"]);
+        }
+
         app(ProductVariantRepositoryInterface::class)->delete($id);
         return response(['status' =>'success', 'message' => 'Product Variant deleted successfully']);
     }
@@ -91,4 +99,5 @@ class ProductVariantController extends Controller
         app(ProductVariantRepositoryInterface::class)->updateStatus($data ,$request->id);
         return response(['status' =>'success' , 'message' =>"Status updated successfully"]);
      }
+
 }
